+++ WebPBL3/DTO/Statistic/StatisticTable.cs
@@ -8,8 +8,8 @@
         public string StaffID { get; set; }
         public DateTime Date { get; set; }
         public int Quantity { get; set; }
-        public int Price { get; set; }
-        public int Total { get; set; }
+        public double Price { get; set; }
+        public double Total { get; set; }
